Refactor/misc (#35)
* refacotor: reorganize files in static folder and remove assets folder in page golder

* refactor: add projects, skills, socials and info to the config file

* refactor: use context hook for avatar and facts, and import DescAboutMe

* refactor: rename props

* refactor: rename props for title component

* refactor: use context hook for socialLinks and add IdealImage

* refactor: use context hook for skills

* refactor: rename props for content title

* refactor: add type check for techs props

* refactor: add one more link for code

* refactor: simplified props by replace links with demolink and codelink

* refactor: use context hook for projects and add code and demo links

* refactor: add type check for Contact component

* refactor: remove props for AboutMe

* refactor: remove projects array and props for Projects component

* styles: run formatter
