+++ docs/learning/data-structure-algo.md
@@ -6,10 +6,12 @@ title: Data Structures and Algorithms
 Everything about Data Structures and Algorithms!
 
 ## Materials
+
 - [JavaScript Algorithms by Codevolution](https://www.youtube.com/playlist?list=PLC3y8-rFHvwjPxNAKvZpdnsr41E0fCMMP)
 - [JavaScript Data Structures by Codevolution](https://www.youtube.com/playlist?list=PLC3y8-rFHvwg6nsAOfC5Is18KB2DrVOJy)
 
 ## Where to practice
+
 - [codewars](https://www.codewars.com/)
 - [LeetCode](https://leetcode.com/)
 - [HackerRank](https://www.hackerrank.com/)
