+++ src/components/AboutMe/index.tsx
@@ -3,10 +3,15 @@ import clsx from "clsx";
 import Image from "@theme/IdealImage";
 import List from "../List";
 import SectionTitle from "../SectionTitle";
-import { facts } from "./info";
 import styles from "./index.module.css";
+import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+import DescAboutMe from "./_about-me.md";
 
-function About({ avatar, AboutMeDesc }) {
+function About() {
+  const { siteConfig } = useDocusaurusContext();
+  const { customFields } = siteConfig;
+  const { avatar, facts } = customFields;
+
   return (
     <section className="container container--fluid">
       <SectionTitle title="About Me" />
@@ -14,11 +19,13 @@ function About({ avatar, AboutMeDesc }) {
         <div className={styles.avatar}>
           <Image img={avatar} alt="Victoria" />
         </div>
-        <section className={styles.para}>{AboutMeDesc}</section>
+        <section className={styles.para}>
+          <DescAboutMe />
+        </section>
       </div>
       <h3>Random or not so random facts about myself:</h3>
       <List classes={styles.facts__list}>
-        {facts.map((fact) => (
+        {(facts as string[]).map((fact) => (
           <li key={fact}>{fact}</li>
         ))}
       </List>
