--- src/components/AboutMe/info.ts
@@ -1,6 +1,0 @@
-export const facts: string[] = [
-  "Deaf",
-  "Gamer",
-  "Love ☕ coffee and 🍵 tea",
-  "Known 3 languages - Mandarin, English, and American Sign Language",
-];
