+++ src/components/List/index.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import clsx from "clsx";
 import styles from "./index.module.css";
 
-interface Props {
+interface ListProps {
   children: React.ReactNode;
   classes?: string;
 }
 
-function List({ children, classes }: Props) {
+function List({ children, classes }: ListProps) {
   return <ul className={clsx(classes, styles.list)}>{children}</ul>;
 }
