+++ src/components/ProjectCard/ContentTitle.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import styles from "./index.module.css";
 
-interface Props {
+interface ContentTitleProps {
   title: string;
 }
 
-function ContentTitle({ title }: Props) {
+function ContentTitle({ title }: ContentTitleProps) {
   return <h4 className={styles.content__title}>{title}</h4>;
 }
