+++ src/components/ProjectCard/ImageContainer.tsx
@@ -4,23 +4,23 @@ import Image from "@theme/IdealImage";
 import LinkBtn from "./LinkBtn";
 import styles from "./index.module.css";
 
-interface Props {
+interface ImageContainerProps {
   image: string;
   title: string;
-  link: string;
-  name: string;
-  link2: string;
-  name2: string;
+  demoLink: string;
+  codeLink: string;
 }
 
-function ImageContainer({ image, title, link, name, link2, name2 }: Props) {
+function ImageContainer({
+  image,
+  title,
+  demoLink,
+  codeLink,
+}: ImageContainerProps) {
   return (
     <div className={clsx("card__image", styles.image)}>
       <Image img={image} alt={title} />
-      <div className={styles.buttons}>
-        <LinkBtn link={link} name={name} />
-        <LinkBtn link={link2} name={name2} />
-      </div>
+      <LinkBtn demoLink={demoLink} codeLink={codeLink} />
     </div>
   );
 }
