+++ src/components/ProjectCard/index.tsx
@@ -4,27 +4,45 @@ import styles from "./index.module.css";
 import ImageContainer from "./ImageContainer";
 import ContentTitle from "./ContentTitle";
 import TechList from "./TechList";
+import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 
-function ProjectCard({ title, image, description, techs, code, demo }) {
+interface Projects {
+  title: string;
+  code: string;
+  demo: string;
+  image: string;
+  description: string;
+  techs: string[];
+}
+
+function ProjectCard() {
+  const { siteConfig } = useDocusaurusContext();
+  const { customFields } = siteConfig;
+  const { projects } = customFields;
+
   return (
-    <div className="col col--4">
-      <section className={clsx("card", styles.card)}>
-        <ImageContainer
-          image={image}
-          title={title}
-          link={code}
-          name="code"
-          link2={demo}
-          name2="demo"
-        />
-        <div className={clsx("card__body")}>
-          <ContentTitle title="Description:" />
-          <p>{description}</p>
-          <ContentTitle title="Tech stacks:" />
-          <TechList techs={techs} />
-        </div>
-      </section>
-    </div>
+    <>
+      {(projects as Projects[]).map(
+        ({ image, title, code, demo, description, techs }) => (
+          <div key={title} className="col col--4">
+            <section className={clsx("card", styles.card)}>
+              <ImageContainer
+                image={image}
+                title={title}
+                codeLink={code}
+                demoLink={demo}
+              />
+              <div className={clsx("card__body")}>
+                <ContentTitle title="Description:" />
+                <p>{description}</p>
+                <ContentTitle title="Tech stacks:" />
+                <TechList techs={techs} />
+              </div>
+            </section>
+          </div>
+        )
+      )}
+    </>
   );
 }
