+++ src/components/SectionTitle/index.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import clsx from "clsx";
 import styles from "./index.module.css";
 
-interface Props {
+interface TitleProps {
   title: string;
 }
 
-function SectionTitle({ title }: Props) {
+function SectionTitle({ title }: TitleProps) {
   return (
     <h2 className={clsx("padding-bottom--md", styles.heading)}>{title}</h2>
   );
