+++ src/components/SocialLinks/index.tsx
@@ -1,15 +1,27 @@
 import React from "react";
-import { socialList } from "./socialList";
 import List from "../List";
 import styles from "./index.module.css";
+import Image from "@theme/IdealImage";
+import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+
+type SocialTitles = "linkedin" | "twitter" | "github" | "polywork";
+
+interface SocialLinks {
+  title: SocialTitles;
+  link: string;
+}
 
 function SocialLinks() {
+  const { siteConfig } = useDocusaurusContext();
+  const { customFields } = siteConfig;
+  const { socialLinks } = customFields;
+
   return (
     <List classes={styles.social}>
-      {socialList.map(({ title, link }) => (
+      {(socialLinks as SocialLinks[]).map(({ title, link }) => (
         <li key={title} className={styles.list__link}>
           <a href={link} target="_blank" rel="noopener">
-            <img alt={title} src={`./img/${title}.svg`} />
+            <Image alt={title} img={`./img/socials/${title}.svg`} />
           </a>
         </li>
       ))}
