--- src/components/skills/skillsList.ts
@@ -1,82 +1,0 @@
-interface Skills {
-  title: string;
-  address: string;
-}
-
-export const skillsList: Skills[] = [
-  {
-    title: "TypeScript",
-    address:
-      "https://img.shields.io/badge/TypeScript-3178C6.svg?style=for-the-badge&logo=TypeScript&logoColor=white",
-  },
-  {
-    title: "TypeScript",
-    address:
-      "https://img.shields.io/badge/JavaScript-F7DF1E.svg?style=for-the-badge&logo=JavaScript&logoColor=black",
-  },
-  {
-    title: "Jest",
-    address:
-      "https://img.shields.io/badge/Jest-C21325.svg?style=for-the-badge&logo=Jest&logoColor=white",
-  },
-  {
-    title: "React",
-    address:
-      "https://img.shields.io/badge/React-61DAFB.svg?style=for-the-badge&logo=React&logoColor=black",
-  },
-  {
-    title: "React Router",
-    address:
-      "https://img.shields.io/badge/React%20Router-CA4245.svg?style=for-the-badge&logo=React-Router&logoColor=white",
-  },
-  {
-    title: "GitHub Action",
-    address:
-      "https://img.shields.io/badge/GitHub%20Actions-2088FF.svg?style=for-the-badge&logo=GitHub-Actions&logoColor=white",
-  },
-  {
-    title: "Vite",
-    address:
-      "https://img.shields.io/badge/Vite-646CFF.svg?style=for-the-badge&logo=Vite&logoColor=white",
-  },
-  {
-    title: "Node.js",
-    address:
-      "https://img.shields.io/badge/Node.js-339933.svg?style=for-the-badge&logo=nodedotjs&logoColor=white",
-  },
-  {
-    title: "Express",
-    address:
-      "https://img.shields.io/badge/Express-000000.svg?style=for-the-badge&logo=Express&logoColor=white",
-  },
-  {
-    title: "SASS",
-    address:
-      "https://img.shields.io/badge/Sass-CC6699.svg?style=for-the-badge&logo=Sass&logoColor=white",
-  },
-  {
-    title: "Tailwind CSS",
-    address:
-      "https://img.shields.io/badge/Tailwind%20CSS-06B6D4.svg?style=for-the-badge&logo=Tailwind-CSS&logoColor=white",
-  },
-  {
-    title: "Styled-components",
-    address:
-      "https://img.shields.io/badge/styledcomponents-DB7093.svg?style=for-the-badge&logo=styled-components&logoColor=white",
-  },
-  {
-    title: "MongoDB",
-    address:
-      "https://img.shields.io/badge/MongoDB-47A248.svg?style=for-the-badge&logo=MongoDB&logoColor=white",
-  },
-  {
-    title: "HTML",
-    address:
-      "https://img.shields.io/badge/HTML5-E34F26.svg?style=for-the-badge&logo=HTML5&logoColor=white",
-  },
-  {
-    title: "CSS",
-    address:
-      "https://img.shields.io/badge/CSS3-1572B6.svg?style=for-the-badge&logo=CSS3&logoColor=white",
-  },
-];
